implicit cast double to float removed
Former-commit-id: 04a90152a1120a2ed1b7c931d29ae9579007db52

+++ src/model/ILine.cpp
@@ -95,8 +95,8 @@ MsgHandler::msgStatus ILine::set (const IMessage* msg)
 		}
 		else if (mode == "wa") {
 			fWAMode = true;
-			float x = a * cos(M_PI * b / 180);
-			float y = a * sin(M_PI * b / 180);
+			double x = a * cos(M_PI * b / 180);
+			double y = a * sin(M_PI * b / 180);
 			setPoint( TFloatPoint(x,y) );
 		}
 		newData(true);
